-- fixed a lot of stuff related to printf-style format specifiers -- properly initialize POD structures
git-svn-id: https://svn.rtmpd.com/crtmpserver/trunk@413 01689507-d990-df11-9d7d-0017a411d416

+++ sources/applications/applestreamingclient/include/protocols/variant/messagestructure.h
@@ -225,7 +225,7 @@ do {\
 		} \
 		default: \
 		{ \
-			ASSERT("Unknown status code: %d",(status)); \
+			ASSERT("Unknown status code: %u",(uint32_t)(status)); \
 			ASC_RES_STATUS_DESC(r)=ASC_RES_STATUS_UNKNOWN; \
 			break;\
 		} \
