+++ sources/thelib/src/netio/kqueue/inboundnamedpipecarrier.cpp
@@ -36,8 +36,8 @@ InboundNamedPipeCarrier *InboundNamedPipeCarrier::Create(string path,
 		uint16_t mode) {
 	if (mkfifo(STR(path), mode) != 0) {
 		int err = errno;
-		FATAL("Unable to create named pipe %s with mode %d: %s (%d)",
-				STR(path), mode, strerror(err), err);
+		FATAL("Unable to create named pipe %s with mode %u: %s (%d)",
+				STR(path), (uint32_t) mode, strerror(err), err);
 		return NULL;
 	}
 
@@ -87,8 +87,7 @@ bool InboundNamedPipeCarrier::OnEvent(struct kevent &event) {
 		}
 		default:
 		{
-			ASSERT("Invalid state: %d", event.filter);
-
+			ASSERT("Invalid state: %hu", event.filter);
 			return false;
 		}
 	}
